refactor(dom): extract DomBoundary from DomModel
